requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.1
django-allauth==0.43.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.2
urllib3==2.0.2
whitenoise==6.4.0
